Split token validation and authorization middleware
Also remove unused superadmin role lookup.

Fix #2049
