+++ internal/controlplane/handlers_authz.go
@@ -16,20 +16,12 @@ package controlplane
 
 import (
 	"context"
-	"fmt"
-	"strings"
 
 	"github.com/google/uuid"
-	gauth "github.com/grpc-ecosystem/go-grpc-middleware/v2/interceptors/auth"
-	"github.com/lestrrat-go/jwx/v2/jwt/openid"
-	"github.com/rs/zerolog"
 	"golang.org/x/exp/slices"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
-	"google.golang.org/protobuf/proto"
-	"google.golang.org/protobuf/reflect/protoreflect"
-	"google.golang.org/protobuf/reflect/protoregistry"
 
 	"github.com/stacklok/minder/internal/auth"
 	"github.com/stacklok/minder/internal/db"
@@ -38,10 +30,6 @@ import (
 )
 
 type rpcOptionsKey struct{}
-
-func withRpcOptions(ctx context.Context, opts *minder.RpcOptions) context.Context {
-	return context.WithValue(ctx, rpcOptionsKey{}, opts)
-}
 
 func getRpcOptions(ctx context.Context) *minder.RpcOptions {
 	// nil value default is okay here
@@ -54,35 +42,22 @@ func isSuperadmin(claims auth.UserPermissions) bool {
 	return claims.IsStaff
 }
 
-func containsSuperadminRole(openIdToken openid.Token) bool {
-	if realmAccess, ok := openIdToken.Get("realm_access"); ok {
-		if realms, ok := realmAccess.(map[string]interface{}); ok {
-			if roles, ok := realms["roles"]; ok {
-				if userRoles, ok := roles.([]interface{}); ok {
-					if slices.Contains(userRoles, "superadmin") {
-						return true
-					}
-				}
-			}
-		}
-	}
-	return false
-}
-
 // lookupUserPermissions returns the user permissions from the database for the given user
-func lookupUserPermissions(ctx context.Context, store db.Store, tok openid.Token) (auth.UserPermissions, error) {
+func lookupUserPermissions(ctx context.Context, store db.Store) auth.UserPermissions {
 	emptyPermissions := auth.UserPermissions{}
 
+	subject := auth.GetUserSubjectFromContext(ctx)
+
 	// read all information for user claims
-	userInfo, err := store.GetUserBySubject(ctx, tok.Subject())
+	userInfo, err := store.GetUserBySubject(ctx, subject)
 	if err != nil {
-		return emptyPermissions, fmt.Errorf("failed to read user")
+		return emptyPermissions
 	}
 
 	// read projects and add id to claims
 	gs, err := store.GetUserProjects(ctx, userInfo.ID)
 	if err != nil {
-		return emptyPermissions, fmt.Errorf("failed to get projects")
+		return emptyPermissions
 	}
 	var projects []uuid.UUID
 	for _, g := range gs {
@@ -92,7 +67,7 @@ func lookupUserPermissions(ctx context.Context, store db.Store, tok openid.Token
 	// read roles and add details to claims
 	rs, err := store.GetUserRoles(ctx, userInfo.ID)
 	if err != nil {
-		return emptyPermissions, fmt.Errorf("failed to get roles")
+		return emptyPermissions
 	}
 
 	var roles []auth.RoleInfo
@@ -114,10 +89,9 @@ func lookupUserPermissions(ctx context.Context, store db.Store, tok openid.Token
 		Roles:          roles,
 		ProjectIds:     projects,
 		OrganizationId: userInfo.OrganizationID,
-		IsStaff:        containsSuperadminRole(tok),
 	}
 
-	return claims, nil
+	return claims
 }
 
 // AuthorizedOnProject checks if the request is authorized for the given
@@ -148,62 +122,16 @@ func AuthorizedOnProject(ctx context.Context, projectID uuid.UUID) error {
 	return nil
 }
 
-// AuthUnaryInterceptor is a server interceptor for authentication
-func AuthUnaryInterceptor(ctx context.Context, req interface{}, info *grpc.UnaryServerInfo,
+// AuthorizationUnaryInterceptor is a server interceptor that sets up the user permissions
+func AuthorizationUnaryInterceptor(ctx context.Context, req interface{}, info *grpc.UnaryServerInfo,
 	handler grpc.UnaryHandler) (any, error) {
-
-	opts, err := optionsForMethod(info)
-	if err != nil {
-		// Fail closed safely, rather than log and proceed.
-		return nil, status.Errorf(codes.Internal, "Error getting options for method: %v", err)
-	}
-
-	ctx = withRpcOptions(ctx, opts)
-
-	if opts.GetAnonymous() {
-		if !opts.GetNoLog() {
-			zerolog.Ctx(ctx).Info().Msgf("Bypassing authentication")
-		}
-		return handler(ctx, req)
-	}
-
-	token, err := gauth.AuthFromMD(ctx, "bearer")
-	if err != nil {
-		if statusErr, ok := status.FromError(err); ok {
-			return nil, util.FromRpcError(statusErr)
-		}
-		return nil, status.Errorf(codes.Unauthenticated, "no auth token: %v", err)
-	}
 
 	server := info.Server.(*Server)
 
-	parsedToken, err := server.vldtr.ParseAndValidate(token)
-	if err != nil {
-		return nil, status.Errorf(codes.Unauthenticated, "invalid auth token: %v", err)
-	}
-
 	// get user authorities from the database
 	// ignore any error because the user may not exist yet
-	authorities, _ := lookupUserPermissions(ctx, server.store, parsedToken)
-
-	if opts.GetRootAdminOnly() && !isSuperadmin(authorities) {
-		return nil, status.Errorf(codes.PermissionDenied, "user not authorized")
-	}
+	authorities := lookupUserPermissions(ctx, server.store)
 
 	ctx = auth.WithPermissionsContext(ctx, authorities)
 	return handler(ctx, req)
-}
-
-func optionsForMethod(info *grpc.UnaryServerInfo) (*minder.RpcOptions, error) {
-	formattedName := strings.ReplaceAll(info.FullMethod[1:], "/", ".")
-	descriptor, err := protoregistry.GlobalFiles.FindDescriptorByName(protoreflect.FullName(formattedName))
-	if err != nil {
-		return nil, fmt.Errorf("unable to find descriptor for %q: %w", formattedName, err)
-	}
-	extension := proto.GetExtension(descriptor.Options(), minder.E_RpcOptions)
-	opts, ok := extension.(*minder.RpcOptions)
-	if !ok {
-		return nil, fmt.Errorf("couldn't decode option for %q, wrong type: %T", formattedName, extension)
-	}
-	return opts, nil
 }
